D3ASIM-3589: Ported SettlementMarket to its own module. Cosmetic fixes.
