+++ src/d3a/models/area/__init__.py
@@ -19,8 +19,15 @@ from logging import getLogger
 from typing import List  # noqa
 from uuid import uuid4
 
-import d3a.constants
 from cached_property import cached_property
+from d3a_interface.area_validator import validate_area
+from d3a_interface.constants_limits import ConstSettings, GlobalConfig
+from d3a_interface.enums import SpotMarketTypeEnum
+from d3a_interface.utils import key_in_dict_and_not_none
+from pendulum import DateTime, duration, today
+from slugify import slugify
+
+import d3a.constants
 from d3a.d3a_core.blockchain_interface import blockchain_interface_factory
 from d3a.d3a_core.device_registry import DeviceRegistry
 from d3a.d3a_core.exceptions import AreaException
@@ -34,14 +41,9 @@ from d3a.models.area.redis_external_market_connection import RedisMarketExternal
 from d3a.models.area.stats import AreaStats
 from d3a.models.area.throughput_parameters import ThroughputParameters
 from d3a.models.config import SimulationConfig
+from d3a.models.market.market_structures import MarketClassType
 from d3a.models.strategy import BaseStrategy
 from d3a.models.strategy.external_strategies import ExternalMixin
-from d3a_interface.area_validator import validate_area
-from d3a_interface.constants_limits import ConstSettings, GlobalConfig
-from d3a_interface.enums import SpotMarketTypeEnum
-from d3a_interface.utils import key_in_dict_and_not_none
-from pendulum import DateTime, duration, today
-from slugify import slugify
 
 log = getLogger(__name__)
 
@@ -349,7 +351,9 @@ class Area:
             self._update_descendants_strategy_prices()
             self.should_update_child_strategies = False
 
-        from d3a.models.area.market_rotators import MarketClassType
+        # TODO: Refactor and port the future, spot, settlement and balancing market creation to
+        # AreaMarkets class, in order to create all necessary markets with one call.
+
         # Markets range from one slot to market_count into the future
         changed = self._markets.create_future_markets(now_value, MarketClassType.SPOT, self)
